// backend/src/repositories/auth.repo.d.ts
export declare const findUserByEmail: (email: string) => Promise<{
    id: number;
    name: string;
    email: string;
    password: string;
    role: "CANDIDATE";
    userType: "candidato";
    isActive: boolean;
} | {
    id: number;
    name: string;
    email: string;
    password: string;
    role: "COMPANY";
    userType: "empresa";
    isActive: boolean;
} | {
    id: number;
    name: string;
    email: string;
    password: string;
    role: "ADMIN";
    userType: "admin";
    isActive: boolean;
} | null>;
export declare const createCandidato: (data: {
    name: string;
    email: string;
    password: string;
}) => Promise<{
    id: number;
    nome: string;
    createdAt: Date;
    updatedAt: Date;
    escolaridade: string | null;
    isActive: boolean;
    email: string | null;
    cpf: string | null;
    telefone: string | null;
    profileData: import("@prisma/client/runtime/library").JsonValue | null;
    curriculoUrl: string | null;
    laudoMedicoUrl: string | null;
    cep: string | null;
    cidade: string | null;
    dataNascimento: Date | null;
    endereco: string | null;
    estado: string | null;
    genero: string | null;
    linkedin: string | null;
    password: string | null;
    portfolio: string | null;
    avatarUrl: string | null;
    experiencias: string | null;
}>;
export declare const createEmpresa: (data: {
    name: string;
    email: string;
    password: string;
}) => Promise<{
    id: number;
    nome: string;
    descricao: string | null;
    createdAt: Date;
    updatedAt: Date;
    barreiras: string | null;
    isActive: boolean;
    email: string | null;
    telefone: string | null;
    cep: string | null;
    cidade: string | null;
    endereco: string | null;
    estado: string | null;
    password: string | null;
    cnpj: string | null;
    companyData: import("@prisma/client/runtime/library").JsonValue | null;
    nomeFantasia: string | null;
    porte: string | null;
    razaoSocial: string | null;
    setor: string | null;
    site: string | null;
    acessibilidadesOferecidas: string | null;
    bairro: string | null;
    complemento: string | null;
    logradouro: string | null;
    numero: string | null;
    outrasBarreiras: string | null;
    outrosRecursosAcessibilidade: string | null;
    politicasInclusao: string | null;
    possuiSistemaInterno: boolean | null;
    responsavelCargo: string | null;
    responsavelEmail: string | null;
    responsavelNome: string | null;
    responsavelTelefone: string | null;
}>;
export declare const createAdmin: (data: {
    name: string;
    email: string;
    password: string;
}) => Promise<{
    id: number;
    nome: string;
    createdAt: Date;
    updatedAt: Date;
    isActive: boolean;
    email: string;
    password: string;
    cargo: string;
    permissoes: import("@prisma/client/runtime/library").JsonValue | null;
}>;